+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/InAppWebView/InAppWebViewClient.java
@@ -179,6 +179,10 @@ public class InAppWebViewClient extends WebViewClient {
     if (webView.options.useOnLoadResource) {
       js += InAppWebView.resourceObserverJS.replaceAll("[\r\n]+", "");
     }
+    js += InAppWebView.checkGlobalKeyDownEventToHideContextMenuJS.replaceAll("[\r\n]+", "");
+    if (flutterWebView != null) {
+      js += InAppWebView.androidKeyboardWorkaroundFocusoutEventJS.replaceAll("[\r\n]+", "");
+    }
     js += InAppWebView.printJS.replaceAll("[\r\n]+", "");
 
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {